refactor: 상위 패키지 이름 `user` -> `member`로 변경
