+++ src/main/java/com/wanted/domain/member/dao/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.user.dao;
+package com.wanted.domain.member.dao;
 
-import com.wanted.domain.user.entity.Member;
+import com.wanted.domain.member.entity.Member;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
