+++ src/main/java/com/wanted/domain/member/dao/location/MemberLocationRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.user.dao.location;
+package com.wanted.domain.member.dao.location;
 
-import com.wanted.domain.user.entity.location.MemberLocation;
+import com.wanted.domain.member.entity.location.MemberLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberLocationRepository extends JpaRepository<MemberLocation, Long> {
